Tl/firestore query (#208)
Adds the collection-filter functionality to the CollectionList... specifically filtering Documents that have certain field-attributes.

The main gist:

Added a store to Firestore, with actions/selectors
The store maintains a CollectionFilter for a given CollectionPath for the life of the component
The panel is really just a fancy-form that tries to keep everything typesafe (not super easy with react-hook-form); and reads/writes to the store
The Collection component conditionally adds where and orderBy clauses to the query based on values in the store
The UI mostly stays true to the Console with the minor caveat that Console is essentially an accordion that only shows one panel at a time, this implementation shows any/all panels. If we get push back, we can add that functionality, but it will require some TLC for keyboard navigation/auto-focus.