Firestore re-arch w/ hooks+suspense (#345)
At a high-level, this is a re-architecture of the Firestore feature to rely on hooks+suspense. The biggest benefit here is that since we have nested-routing (where a document conditionally renders a sub-collection, and a collection conditionally renders a sub-document), we can now render as many documents/collections as are in the ROUTE and then lazily populate data in each component as it comes in over the wire; this gets us "missing document" support.