PR-Firestore-Requests-Card (#477)
This PR focuses on adding the RequestsCard components, which are divided into 2 folders: Header (filters header) and Table.

The Header folder contains an empty component for now, since the filters are not yet developed, and will only be developed as a stretch goal if there is time left.

The Table folder contains the Table, a ZeroState, and a NoResults components. Also, these components already expect the data to be filtered when received as props, so no modification will have to be made to this folder after the filters from the Header folder are developed.