PR-Firestore-Request-Details-Code-Viewer (#488)
This is the 8th PR created to review the code of my intern project.

This PR focuses on adding the Code Viewer section of the RequestDetails view.

Additionally, I restructured the RequestDetails folder by splitting the files into 3 sub-folders (Header, CodeViewer, InspectionSection) and an index file. I took this decision because the RequestDetails folder was getting very big.