Firestore Requests (#602)
* Firestore Rules Store and Evaluations definitions. (#453)

* Changed the firestore tabs routing approach to avoid using the global routes array, instead used a nested routing approach (#475)

* Firestore request evaluations global subscription (#476)

Added global subscription to request evaluations websocket, and mocked 'admin' and 'error' request evaluations by injecting mock data to incoming evaluations

* PR-Firestore-Requests-Card (#477)

This PR focuses on adding the RequestsCard components, which are divided into 2 folders: Header (filters header) and Table.

The Header folder contains an empty component for now, since the filters are not yet developed, and will only be developed as a stretch goal if there is time left.

The Table folder contains the Table, a ZeroState, and a NoResults components. Also, these components already expect the data to be filtered when received as props, so no modification will have to be made to this folder after the filters from the Header folder are developed.

* PR-Firestore-Requests-Table-Row (#478)

This is the 5th PR created to review the code of my intern project.

This PR focuses on adding the RequestsTableRow content (UI, SCSS, and unit test).

Additionally, the 6th PR (PR-Firestore-Request-Details #479) was merge into this one.

* PR-Firestore-Request-Details-Header (#484)

This is the 7th PR created to review the code of my intern project.

This PR focuses on adding the Header of the RequestDetails view.

Additionally, I fixed 2 flaky tests that were missing asynchronous queries.

* PR-Firestore-Request-Details-Code-Viewer (#488)

This is the 8th PR created to review the code of my intern project.

This PR focuses on adding the Code Viewer section of the RequestDetails view.

Additionally, I restructured the RequestDetails folder by splitting the files into 3 sub-folders (Header, CodeViewer, InspectionSection) and an index file. I took this decision because the RequestDetails folder was getting very big.

* PR-Firestore-Request-Details-Inspection-Section (#489)

This is the 9th PR created to review the code of my intern project.

This PR focuses on adding the inspection section of the RequestDetails view.

* PR-Firestore-Request-Path (#493)

This is the 10th PR created to review the code of my intern project.

This PR focuses on adding the following components:

RequestPath: responsible for rendering the request-path, rendering the copy-icon-button, truncating the request-path (from the left side) and copying to clipboard the request-path after the copy-icon-button is clicked.
CopyPathNotification: responsible for rendering the Snackbar (notification) that appears after the copy-icon-button is clicked.

* Removed actions-buttons from details header

* Removed query-mocked-data to inspect

* Removed sorting arrow from requests-table because sorting functionality is not yet implemented

* Minimized the amount of mocked data injected to new evaluations

* Update for new data format

* Migrate Firestore requests off Redux.

* Fix field previews

* Fix empty object error; remove console.logs

* Add callout for admin requests and access calls.

* Fix banner layout.

* Only show requests from the current project.

* chore(deps): update dependency postcss to 7.0.36 [security] (#594)

Co-authored-by: Renovate Bot <bot@renovateapp.com>

* chore(deps): update firebase packages (#593)

Co-authored-by: Renovate Bot <bot@renovateapp.com>

* chore(release): 1.6.0

* Add requests.auth.

* Add request.query and fix timestamps. (#599)

* Add a callout for no matching requests; fix admin request callout (#598)

* Update callout layouts; add no-matching callout

* Go back to readonly mode

* Fix some styles

* PR feedback

* Make CI run on PRs to all branches.

* Deflake some tests.

* Switch to public stable firebase-tools.

* Fix waitForElementToBeRemoved usage.

* Fix a few leaky tests (overlapping act).

* Add some valid data for firestore request tests.

* Fix more tests.

* Disable gutter tests for now.

* Run prettier.

* Improve syntax highlighting. (#600)

* chore(deps): update dependency postcss to 7.0.36 [security] (#594)

Co-authored-by: Renovate Bot <bot@renovateapp.com>

* chore(deps): update firebase packages (#593)

Co-authored-by: Renovate Bot <bot@renovateapp.com>

* Improve syntax highlighting.

* Apply suggestions from code review

Co-authored-by: Sam <samgho@google.com>

* Format files using prettier.

* Fix test on rules code viewer.

* Add some valid data for firestore request tests.

* Deflake some tests.

* Switch to public stable firebase-tools.

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>
Co-authored-by: Renovate Bot <bot@renovateapp.com>
Co-authored-by: Sam <samgho@google.com>

* Refactor FieldPreview and handle rules types. (#606)

* Refactor FieldPreview and handle rules types.

* Update tests.

* Add loading suspense for requests. (#609)

* Remove one extra $ symbol.

* Ys/requests suspend fix (#613)

* Fix FirestoreRequestProvider loop.

* Add fallback for requests unsupported.

* Fix boolean typing.

* Fix tests.

* Fix boolean.

Co-authored-by: Antonio Rodriguez <rodantonio@google.com>
Co-authored-by: Sam Olsen <samgho@google.com>
Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>
Co-authored-by: Renovate Bot <bot@renovateapp.com>