Fix windows issue with npm start
Use the run-script-os npm package to detect if the OS is Windows when running npm start. If so, it runs npm start:win32. Otherwise, it runs npm run start:default.

Also, instead of using the dotenv npm package to set the SASS_PATH variable in the .env file, now it is being set inline, in the npm scripts start:win32 and start:default.

Fix #210.