Fix issue in the build:web script, caused by the previous commit
In the previous commit (53224), the environment variable SASS_PATH was removed from the .env file and added only to the start npm scripts (start:win32 and start:default).
However, the build:web script also needs that environment variable.
This commit sets that SASS_PATH for build:web, in the same fashion as it was set for the start scripts, using the run-script-os package.