Fix windows issue with npm start (#637)
* Fix windows issue with npm start

Use the run-script-os npm package to detect if the OS is Windows when running npm start. If so, it runs npm start:win32. Otherwise, it runs npm run start:default.

Also, instead of using the dotenv npm package to set the SASS_PATH variable in the .env file, now it is being set inline, in the npm scripts start:win32 and start:default.

Fix #210.

* Fix issue in the build:web script, caused by the previous commit

In the previous commit (53224), the environment variable SASS_PATH was removed from the .env file and added only to the start npm scripts (start:win32 and start:default).
However, the build:web script also needs that environment variable.
This commit sets that SASS_PATH for build:web, in the same fashion as it was set for the start scripts, using the run-script-os package.

Co-authored-by: Yuchen Shi <yuchenshi@google.com>