Fix issue with the order of requests in the Firestore tab
By using the root time field, instead of the rulesContext.time field.

The former seems to have the timestamp at which the request was actually sent. On the other hand, the latter apparently keeps the timestamp at which some web socket communication was created. In the case of get-only-once firestore requests, that connection seems to last for 1 minute,

I had to add the time field to the FirestoreRulesEvaluation interface, as it wasn't there. And also, add the time field to the object returned by a testing util function.

Fix #634.