fix(a11y) Show focus/hover shadow/ipple effect on BreadCrumb Buttons
This commit adds back the ripple effect for buttons within the
Breadcrumb area.

This fixes the issue where the "Import JSON" button was not accessible.
In truth, the button was always accessible, but without the ripple/shadow
the user doesn't know they are focused on the button.

FIXED=259452701, b/259452701
b/259452701