fix(a11y) Fix double-focus upload-button issue (#879)
* fix(a11y) Fix double-focus upload-button issue

This commit fixes an issue where the "Upload File" section of the
Storage page contained 2 click targets.

It turns out the 2 click targets were the div wrapper and the button
itself.

For a11y, we want the focus target to be on the button instead of the
div (the screen reader will announce the user is focused on a button
when reading out the aria-label.)

FIXED=259449474, b/259449474
b/259449474

* fix lint issues with `npm run lint:fix`