fix(a11y) Recapture focus after closing ImportJson dialog
This commit fixes an issue where the focus is suddenly lost when the
user closes the "Import JSON" on the RTDB page.

Previously, when the dialog closes the focus goes to the bottom of the
page (or resets to the head.)

With this commit, we recapture the focus and focus on the "More options"
menu. It would have been better to put the focus back on the
"Import JSON" button... but the problem is the "Import JSON" button is
a child of menu that disappears and becomes un-focusable once selected.

FIXED=259453177, b/259453177

b/259453177