fix(a11y) Update "IMPORT JSON" MenuItem to respond to keyboard-only
This commit updates the SimpleMenu and the MenuItem associated with the
"Import Json" MenuItem such that you can activate/select the "Import
Json" dialog using the keyboard.

The problem before was pressing "Enter" or "Space" would not activate
the "onClick" handler.

But, it would bubble up a "selected" event.

So instead I moved the "selected" behavior into the Menu and am using
a dataset attribute to control behavior.

FIXED=259452701, b/259452701